+++ src/components/IndexPage/mission.js
@@ -2,8 +2,8 @@ import React from "react"
 import PropTypes from "prop-types"
 import styled from "styled-components"
 
-import { Header, HeaderLineBelow } from "../styles/Header"
-import Section from "../styles/Section"
+import { Header, HeaderLineBelow } from "styles/Header"
+import Section from "styles/Section"
 
 const BlueBackground = styled.div`
   background: #2a426b;
